tools/linkedin_tools.py
from crewai_tools import BaseTool
import requests
import csv

class FetchLinkedInTrendsTool(BaseTool):
    name = "fetch_linkedin_trends"
    description = "Fetches trending LinkedIn posts or topics based on user topic."

    def _run(self, topic: str):
        # Replace with real API or scraping logic
        return f"Sample trends related to {topic}: [AI in radiology, AI for diagnosis]"

class SavePostTool(BaseTool):
    name = "save_post_to_file"
    description = "Saves the LinkedIn post to a text file."

    def _run(self, post: str):
        with open("linkedin_post.txt", "w", encoding="utf-8") as file:
            file.write(post)
        return "Post saved successfully."

class PostToLinkedInTool(BaseTool):
    name = "post_to_linkedin"
    description = "Posts content to LinkedIn using the LinkedIn API."

    def _run(self, post: str):
        # You must integrate LinkedIn OAuth + API here
        # Placeholder response
        return f"Post sent to LinkedIn: {post[:100]}..."
