pyproject.toml
[project]
name = "linkedin-post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.150.0",
    "crewai-tools>=0.60.0",
    "ipykernel>=6.30.0",
    "python-dotenv>=0.9.0",
    "pyyaml>=6.0.2",
]
